// frontend/src/api/generated/models/integrationConnectionStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type IntegrationConnectionStatus = typeof IntegrationConnectionStatus[keyof typeof IntegrationConnectionStatus];


export const IntegrationConnectionStatus = {
  NotConnected: 'NotConnected',
  Connected: 'Connected',
  Error: 'Error',
  Revoked: 'Revoked',
} as const;