// frontend/src/api/generated/models/integrationProviderType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type IntegrationProviderType = typeof IntegrationProviderType[keyof typeof IntegrationProviderType];


export const IntegrationProviderType = {
  Strava: 'Strava',
  GarminDemoProvider: 'GarminDemoProvider',
  MySasyDemoProvider: 'MySasyDemoProvider',
} as const;